Caching ArrayFire and Cmake

+++ .CI/travis/install-arrayfire_linux.sh
@@ -5,15 +5,13 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+if [ ! -d "/opt/arrayfire" ]; then
+    # Control will enter here if $DIRECTORY doesn't exist.
+    wget http://arrayfire.s3.amazonaws.com/3.5.1/ArrayFire-v3.5.1_Linux_x86_64.sh
+    sudo mkdir /opt -p
 
-#if [ ! -d "/opt/arrayfire" ]; then
-  # Control will enter here if $DIRECTORY doesn't exist.
-  wget http://arrayfire.s3.amazonaws.com/3.5.1/ArrayFire-v3.5.1_Linux_x86_64.sh
-  sudo mkdir /opt -p
+    chmod +x ArrayFire-v3.5.1_Linux_x86_64.sh
+    sudo ./ArrayFire-v3.5.1_Linux_x86_64.sh --prefix=/opt --skip-license
 
-  chmod +x ArrayFire-v3.5.1_Linux_x86_64.sh
-  sudo ./ArrayFire-v3.5.1_Linux_x86_64.sh --prefix=/opt --skip-license
-
-  rm ArrayFire-v3.5.1_Linux_x86_64.sh
-#fi
-
+    rm ArrayFire-v3.5.1_Linux_x86_64.sh
+fi
